tests/wpt/tests/html/canvas/element/drawing-rectangles-to-the-canvas/2d.strokeRect.transform.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.strokeRect.transform</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.strokeRect.transform</h1>
<p class="desc">fillRect is affected by transforms</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("fillRect is affected by transforms");
_addTest(function(canvas, ctx) {

  ctx.scale(10, 10);
  ctx.translate(0, 5);
  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 5;
  ctx.strokeRect(2.5, -2.6, 5, 0.2);
  _assertPixel(canvas, 50,25, 0,255,0,255);

});
</script>
